add:main page and my nft

+++ contracts/NFT.sol
@@ -18,12 +18,12 @@ import "@openzeppelin/contracts/utils/Counters.sol";
 contract NFT is ERC721URIStorage{
     using Counters for Counters.Counter;
     Counters.Counter private _tokenIds;
-    address contractAddress;
+    address contractAddress; 
 
     constructor(address marketPlaceAddress) ERC721("Kavindu","IVAK"){
         contractAddress=marketPlaceAddress;
     }
-    function createToken(string memory tokenURI) public returns(uint256){
+    function createToken(string memory tokenURI) public returns(uint){
         _tokenIds.increment();
         uint256 newItemId=_tokenIds.current();
         _mint(msg.sender,newItemId);
